requirements.txt
# Dependencies
python-telegram-bot>=21.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tzdata>=2024.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
